feat(smartem): purge mocks from live-API mode, add ErrorBoundary coverage
In `dev:smartem` (live API) mode the SPA should render real data or
nothing - never mock data masquerading as real. This was not the case
on the dashboard `/` route, where stale mock instruments and sessions
rendered unconditionally with links to mock acquisition UUIDs that
return empty lists from the real API.

Three rules now hold for the live build:

1. No mock data is reachable at runtime. Every gate uses an inline
   `import.meta.env.VITE_ENABLE_MOCKS === 'true'` check rather than a
   re-export, so Vite parse-time substitution constant-folds the dead
   branch and Rollup tree-shakes the gated imports.
2. Mock-only components are lazy-loaded sibling chunks (MockDashboard,
   MockContextStrip), only fetched when the env flag is on.
3. Where a real-API path is not yet wired, the SPA renders a visible
   placeholder (`EmptyState`) rather than fake data, so gaps are
   loudly visible during the mock-to-real migration.

Added a generic `<ErrorBoundary>` and per-route `errorComponent` so a
broken panel in one route does not bring down the rest of the SPA -
useful while the wiring is still in flux.

Dashboard `/` becomes EmptyState in live mode. Header drops the
mock-fed acquisitions command-palette group (it needs real-API search
to be useful). ViewSwitcher stops looking up grid names from mocks.
ContextStrip splits into a live IDs-only variant and a lazy
MockContextStrip for the rich thumbnails.
